// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      catalogo: {
        Row: {
          categoria: string | null
          created_at: string | null
          deleted_at: string | null
          descricao: string | null
          estoque: number | null
          id_catalogo: number
          img_catalogo: string | null
          nm_catalogo: string
          vlr_item: number
        }
        Insert: {
          categoria?: string | null
          created_at?: string | null
          deleted_at?: string | null
          descricao?: string | null
          estoque?: number | null
          id_catalogo?: number
          img_catalogo?: string | null
          nm_catalogo: string
          vlr_item: number
        }
        Update: {
          categoria?: string | null
          created_at?: string | null
          deleted_at?: string | null
          descricao?: string | null
          estoque?: number | null
          id_catalogo?: number
          img_catalogo?: string | null
          nm_catalogo?: string
          vlr_item?: number
        }
        Relationships: []
      }
      vendas: {
        Row: {
          cliente: string | null
          created_at: string | null
          deleted_at: string | null
          forma_pagamento: string | null
          id_catalogo: number | null
          id_venda: number
          quantidade: number | null
          vlr_transaction: number
        }
        Insert: {
          cliente?: string | null
          created_at?: string | null
          deleted_at?: string | null
          forma_pagamento?: string | null
          id_catalogo?: number | null
          id_venda?: number
          quantidade?: number | null
          vlr_transaction: number
        }
        Update: {
          cliente?: string | null
          created_at?: string | null
          deleted_at?: string | null
          forma_pagamento?: string | null
          id_catalogo?: number | null
          id_venda?: number
          quantidade?: number | null
          vlr_transaction?: number
        }
        Relationships: [
          {
            foreignKeyName: "vendas_id_catalogo_fkey"
            columns: ["id_catalogo"]
            isOneToOne: false
            referencedRelation: "catalogo"
            referencedColumns: ["id_catalogo"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never